[SWM-91] Rename : 'hook' directory to 'hooks'
