+++ src/components/nav/NavBar.tsx
@@ -1,10 +1,8 @@
 'use client';
 import Link from 'next/link';
 import SearchInput from './SearchInput';
-import useAuth from '@/src/hook/useAuth';
+import useAuth from '@/src/hooks/useAuth';
 import Button from '../ui/Button';
-import { signOut } from 'next-auth/react';
-import useAlert from '@/src/hook/useAlert';
 
 type Props = {
   logo: string;
@@ -12,21 +10,10 @@ type Props = {
 };
 
 export default function NavBar({ logo, profileDropdown }: Props) {
-  const { session } = useAuth();
-  const { setAlert } = useAlert();
+  const { session, logout } = useAuth();
 
   const name = session?.name;
   const bio = session?.bio;
-
-  const logoutErrorHandler = async () => {
-    await signOut().catch(() =>
-      setAlert({
-        type: 'error',
-        title: '에러 발생!',
-        description: '로그아웃에 실패했어요'
-      })
-    );
-  };
 
   return (
     <nav className='flex justify-between px-10 navbar'>
@@ -40,7 +27,7 @@ export default function NavBar({ logo, profileDropdown }: Props) {
       </div>
       <div className='flex gap-4'>
         <Button
-          onClick={logoutErrorHandler}
+          onClick={logout}
           className={`${name ? '' : 'hidden'} g_id_signout`}
         >
           <p>로그아웃</p>
